+++ back-end/src/main/java/io/thescenery/shopAssist/user/service/IUserService.java
@@ -5,7 +5,7 @@ import io.thescenery.shopAssist.user.vo.UserInfoVo;
 
 public interface IUserService {
 
-  public UserInfoVo getUserById(int id);
+  public UserInfoVo getUserById(long id);
 
   public User getUserByEmail(String email);
 }
